// src/model/fullItemAllOf.ts
/**
 * 1Password Connect
 * API interface for the 1Password Connect server.
 *
 * The version of the OpenAPI document: 0.2.3
 * Contact: support@1password.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { FullItemAllOfFields } from './fullItemAllOfFields';
import { FullItemAllOfSections } from './fullItemAllOfSections';

export class FullItemAllOf {
    'sections'?: Array<FullItemAllOfSections>;
    'fields'?: Array<FullItemAllOfFields>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "sections",
            "baseName": "sections",
            "type": "Array<FullItemAllOfSections>"
        },
        {
            "name": "fields",
            "baseName": "fields",
            "type": "Array<FullItemAllOfFields>"
        }    ];

    static getAttributeTypeMap() {
        return FullItemAllOf.attributeTypeMap;
    }
}